[update] 不要なフォルダ削除
