--- Assets/SoulRunProject/Program/Scripts/BoxCastTest.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4f953d9d94c0074459e07f4da43bf7a7
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
