--- Assets/SoulRunProject/Program/Scripts/BoxCastTest/Core.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 35f0260b62fb5b542a7375fd4ae56cce
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
