--- Assets/SoulRunProject/Program/Scripts/BoxCastTest/Interface.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1b78e3164ae3e9d4aa2a8cf175815dc7
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
