--- Assets/SoulRunProject/Program/Scripts/BoxCastTest/UI.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ddbcf1ca3ec2b774da05d0f496a881c6
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
