--- Assets/SoulRunProject/Program/Scripts/HikanyanDebugScene.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1e0dfc5e05a76884c87d423c52d855d4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
