--- Assets/SoulRunProject/Program/Scripts/HikanyanDebugScene/Core.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ecc547863670b9440818308c71c987cb
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
