--- Assets/SoulRunProject/Program/Scripts/HikanyanDebugScene/Interface.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a02637df92d76ab4ea585bb30f0718da
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
