--- Assets/SoulRunProject/Program/Scripts/HikanyanDebugScene/UI.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a3bd1d02f8827d24ea13d2544f6cb41b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
