--- Assets/SoulRunProject/Program/Scripts/SkeiraTest.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ef0dfffd1f5c75b4fb5a63ddb6081d8b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
