--- Assets/SoulRunProject/Program/Scripts/SkeiraTest/Core.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 71d0ffcf281343b4a9637298b7a662eb
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
