--- Assets/SoulRunProject/Program/Scripts/SkeiraTest/Interface.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 926d1d6a0ada71348a90c2fe5f3f55f8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
