--- Assets/SoulRunProject/Program/Scripts/SkeiraTest/UI.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 72001eeb261a17d489dae2ed5ed94f21
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
