--- Assets/SoulRunProject/Program/Shaders/BoxCastTest.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 418594c3ed94b5646b46f2b32151f725
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
