--- Assets/SoulRunProject/Program/Shaders/Common.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3a72e183b6205be45aa9510677dbe6f1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
