--- Assets/SoulRunProject/Program/Shaders/HikanyanDebugScene.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: beabdd5e9bceb1b48baba0eb04a220a0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
