--- Assets/SoulRunProject/Program/Shaders/InGameTest.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 75b4981bda93efe4eaac706ed9913fea
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
