--- Assets/SoulRunProject/Program/Shaders/SceneA.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 85203a693748c72488109d8ed82f598f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
