--- Assets/SoulRunProject/Program/Shaders/SceneB.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d1fdcf06b8c32e84cbd8d006e86d19de
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
