--- Assets/SoulRunProject/Program/Shaders/SkeiraTest.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 188c73480637a7946aa48f88ce073ad8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
