--- Assets/SoulRunProject/Program/Shaders/SoulMixScene.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d2992c0aa995e694c9e89f04b143bf8f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
